libhelfem: relocate arma TEI helpers to src/general/tei_utils.{h,cpp}
utils.h / utils.cpp still hosted three arma-typed two-electron-integral
helpers -- product_tei, check_tei_symmetry, exchange_tei(arma::mat) --
for the diatomic J/K assembly path. The sole caller is
src/diatomic/basis.cpp's prim_tei / prim_ktei build. Move the
declarations + implementations next to that caller as
src/general/tei_utils.{h,cpp}, following the same pattern as the
libhelfem/src/{chebyshev,lobatto}.h relocation in PR #173.

Effects:

- libhelfem/src/utils.h loses `#include <armadillo>` and the three
  arma-typed decls; the surviving helfem::Matrix `exchange_tei`
  overload used by CoulombExchangeFE.h stays in helfem::utils, so
  callers see no namespace change.
- libhelfem/src/utils.cpp loses the three arma-typed implementations
  (about 70 lines) and stops needing <armadillo> via utils.h.
- src/general/tei_utils.{h,cpp} contain identical implementations,
  same helfem::utils namespace so basis.cpp's `utils::product_tei`
  etc. resolve without renames.
- src/CMakeLists.txt adds tei_utils.cpp to helfem-common.
- src/diatomic/basis.cpp replaces `#include "utils.h"` with
  `#include "tei_utils.h"` -- its only utils:: uses were the three
  moved functions.

libhelfem code (both installed and internal) is now arma-free except
for the ArmaEigen.h bridge shim, which exists by design to service
arma-consuming callers.

Verified: H2 SCF (Rbond=1.4, nela=nelb=1, nelem=5, primbas=4)
converges bit-for-bit to -1.1091077979.
